Hoist entries to base class
This changeset moves the vector of entries from the specialized caches to the base cache class.
This will help refactoring and simplication and allow removing all the searchKernel methods in the subsequent changesets.
In particular this requires breaking out the dependent CachedEntry types and making them standalone types that can be passed as template parameters to the base Cache class.
Private visibility and friends are also removed.