Add a benchmark for first and second moments and sum / sumSquares
In particular this benchmark exhibits that we lose a lot on mean / var due
to the normalization update which disables reduction matching.

The performance ratio between (sum, mean) or (sumSquares, var) is about 4x.
Interestingly, the Max outer fuse followed by Min intra fuse works nicely
and allows reductions to occur.

Lastly, fusing sum and sumSquares only gets an extra 10% over the sum of the
kernels which suggests extra fusion opportunities that need to be weighed in
against reduction mathing.

This has been driven by performance analysis of GroupNormalization where we
should be able to perform 2x better than we currently do if we exploited
multiple kernels.