Add min_distances.py
This commit adds an exmaple where the argmin over a reduced sum is required.
This example is now functional thanks to the previous commit but extra
work is needed to make it into a reasonably performing kernel:
1. for the fused kernel to parallelize properly across blocks we need
grid synchronization. This may be a nice concrete use case @math-fehr
2. for the 1-stage fissioned implementation we need device-wide synchronization
otherwise we will always be limited by running on a single SM
3. the 2-stage fissioned implementation can give us performance today